backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
pyyaml==6.0.2
python-multipart==0.0.20
elasticsearch==8.17.0
apscheduler==3.10.4
reportlab==4.2.5
telethon==1.36.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
